Improve query formatting
